+++ lib/utils/app_utils.dart
@@ -5,8 +5,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:notekar/models/palette.dart';
 
-const appVersion = '6.0.5';
-const appBuildNumber = '35';
+const appVersion = '6.0.6';
+const appBuildNumber = '36';
 const appBuildDate = '2026-08-12';
 
 String formatInstalledDate(String buildDateStr) {